ui: move RevisionsView, IndexView to views module
